+++ src/main/java/com/barlinc/unusual_prehistory/blocks/UnderwaterEggBlock.java
@@ -64,15 +64,10 @@ public class UnderwaterEggBlock extends WaterEggBlockEntity implements SimpleWat
                 Entity entity = hatchedEntity.get().create(level);
                 if (entity instanceof Mob mob) {
                     if (entity instanceof Animal animal) {
-                        animal.setAge(-24000);
-                        animal.setPersistenceRequired();
+                        animal.setBaby(true);
                     }
                     if (entity instanceof PrehistoricAquaticMob prehistoricAquaticMob) {
-                        prehistoricAquaticMob.setAge(-24000);
-                        prehistoricAquaticMob.setPersistenceRequired();
-                    }
-                    if (entity instanceof PrehistoricMob prehistoricMob) {
-                        prehistoricMob.setVariant(random.nextInt(prehistoricMob.getVariantCount()));
+                        prehistoricAquaticMob.setFromEgg(true);
                     }
                     int k = random.nextInt(1, 361);
                     if (placer != null) {