+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2DatapackProvider.java
@@ -7,6 +7,7 @@ import net.minecraft.core.RegistrySetBuilder;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.DatapackBuiltinEntriesProvider;
+import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
@@ -14,7 +15,8 @@ import java.util.concurrent.CompletableFuture;
 public class UP2DatapackProvider extends DatapackBuiltinEntriesProvider {
 
     public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
-            .add(Registries.DAMAGE_TYPE, UP2DamageTypes::bootstrap);
+            .add(Registries.DAMAGE_TYPE, UP2DamageTypes::bootstrap)
+            .add(ForgeRegistries.Keys.BIOME_MODIFIERS, UP2BiomeModifierProvider::bootstrap);
 
     public UP2DatapackProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
         super(output, registries,BUILDER, Set.of("minecraft", UnusualPrehistory2.MOD_ID));